refactor(rpc): move invoke DTOs into module
